chore: introduce grid and pathfinding modules
Moved the Position class from its previous module to a newly established 'grid' module.

This change is part of a larger effort to reorganize our project into a more modular, crate-like structure.

By segregating core concepts such as Position and Sector into contexutal modules, we lay the groundwork for future enhancements that include breaking down the Ryot project into smaller, more manageable crates.

This modular architecture will facilitate independent development and maintenance of features, controlled via feature flags, enhancing our ability to scale and update the project efficiently.
