+++ ryot/src/bevy_ryot/game/elevation.rs
@@ -1,7 +1,7 @@
 use crate::bevy_ryot::map::MapTiles;
 use crate::bevy_ryot::sprites::SPRITE_BASE_SIZE;
 use crate::bevy_ryot::{AppearanceAssets, AppearanceGroup, GameObjectId};
-use crate::position::TilePosition;
+use crate::prelude::*;
 use crate::Layer;
 use bevy::prelude::*;
 use itertools::Itertools;
