+++ ryot/src/bevy_ryot/map.rs
@@ -1,8 +1,9 @@
 use bevy::prelude::{Deref, DerefMut, Reflect, Resource};
 use bevy::utils::HashMap;
 
+use crate::layer::Layer;
 use crate::layer::{BottomLayer, Order, RelativeLayer};
-use crate::{layer::Layer, position::TilePosition};
+use crate::prelude::*;
 
 /// A resource that holds the map tiles and the entities that are drawn on them.
 /// An entity location is represented by the combination of a Layer and a Position.
