+++ ryot/src/bevy_ryot/perspective/trajectory.rs
@@ -5,7 +5,6 @@ use bevy::utils::HashSet;
 use crate::bevy_ryot::perspective::*;
 use crate::bevy_ryot::tile_flags::TileFlags;
 use crate::bevy_ryot::CacheSystems;
-use crate::position::TilePosition;
 
 /// Represents an App that can add one or more `Trajectory` to its systems.
 /// Requires the `Cache<RadialArea, Vec<Vec<TilePosition>>>` resource to be initialized.
