+++ ryot/src/lmdb/generator.rs
@@ -81,7 +81,7 @@ house:id -> {
  */
 use crate::layer::{BottomLayer, RelativeLayer};
 use crate::lmdb::*;
-use crate::position::{Sector, TilePosition};
+use crate::prelude::*;
 use crate::Layer;
 use rand::Rng;
