+++ ryot/src/sprites/tests/mod.rs
@@ -1,2 +1,1 @@
-mod position_test;
 mod sprite_layout_test;
