+++ ryot/tests/sprites/sheet_loading.rs
@@ -1,10 +1,7 @@
 use glam::UVec2;
 use image::RgbaImage;
 use rstest::{fixture, rstest};
-use ryot::prelude::{
-    decompress_sprite_sheet, decompress_sprite_sheets, load_sprite_sheet_image, ContentConfigs,
-    SpriteSheetConfig, SPRITE_SHEET_FOLDER,
-};
+use ryot::prelude::*;
 use std::fs::File;
 use std::io::BufReader;
 use std::path::PathBuf;
