+++ ryot_compass/src/lmdb.rs
@@ -4,16 +4,15 @@ use heed::types::Bytes;
 use log::{debug, warn};
 use ryot::bevy_ryot::drawing::{update, CommandState};
 use ryot::bevy_ryot::lmdb::{
-    compact_map, read_area, reload_visible_area, LmdbCompactor, LmdbPlugin as RyotLmdbPlugin,
+    compact_map, read_area, reload_visible_area, LmdbCompactor, LmdbEnv,
+    LmdbPlugin as RyotLmdbPlugin,
 };
 use ryot::bevy_ryot::map::MapTiles;
 use ryot::bevy_ryot::{AsyncEventApp, GameObjectBundle, InternalContentState};
 use ryot::helpers::execute;
 use ryot::lmdb::*;
-use ryot::position::{Sector, TilePosition};
 use ryot::prelude::drawing::TileComponent;
-use ryot::prelude::lmdb::LmdbEnv;
-use ryot::prelude::{compress, decompress, LoadObjects, Zstd};
+use ryot::prelude::*;
 use ryot::{lmdb, Layer};
 use std::collections::HashMap;
 use std::sync::atomic::Ordering;
